requirements-test.txt
pytest
sqlglot
docker
PyYAML
SQLAlchemy
psycopg2
PyMySQL
cx_Oracle
cryptography
